+++ autots/models/sklearn.py
@@ -1822,9 +1822,10 @@ class WindowRegression(ModelObject):
                 forecast_length=forecast_length,
                 forecast_index=df.index,
                 forecast_columns=df.columns,
-                lower_forecast=lower_forecast,
-                forecast=df,
-                upper_forecast=upper_forecast,
+                # so it's producing float32 but pandas is better with float64
+                lower_forecast=lower_forecast.astype(float),
+                forecast=df.astype(float),
+                upper_forecast=upper_forecast.astype(float),
                 prediction_interval=self.prediction_interval,
                 predict_runtime=predict_runtime,
                 fit_runtime=self.fit_runtime,
